% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require2.R
\name{availableVersionOK}
\alias{availableVersionOK}
\title{Needs VersionOnRepos, versionSpec and inequality columns}
\usage{
availableVersionOK(pkgDT)
}
\arguments{
\item{pkgDT}{A \code{pkgDT} object}
}
\description{
Needs VersionOnRepos, versionSpec and inequality columns
}
