% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{cacheDefaultDir}
\alias{cacheDefaultDir}
\title{The default cache directory for Require Cache}
\usage{
cacheDefaultDir()
}
\value{
The default cache directory
}
\description{
A wrapper around \code{tools::R_user_dir("Require", which = "cache")} that
creates the directory, if it does not exist.
}
