% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgSnapshot.R
\name{checkLibPaths}
\alias{checkLibPaths}
\title{Creates the directories, and adds version number}
\usage{
checkLibPaths(libPaths, ifMissing, exact = FALSE, ...)
}
\description{
Creates the directories, and adds version number
}
