% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRAN.R
\name{chooseCRANmirror2}
\alias{chooseCRANmirror2}
\title{Pass through function for \code{chooseCRANmirror}}
\usage{
chooseCRANmirror2(...)
}
\arguments{
\item{...}{Passed to \code{chooseCRANmirror}}
}
\description{
This is here to allow mocking during unit testing related to \code{chooseCRANmirror}.
}
\keyword{internal}
