% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgSnapshot.R
\name{doLibPaths}
\alias{doLibPaths}
\title{Deals with missing libPaths arg, and takes first}
\usage{
doLibPaths(libPaths, standAlone = FALSE)
}
\description{
Deals with missing libPaths arg, and takes first
}
