% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-helpers.R
\name{.downloadFileMasterMainAuth}
\alias{.downloadFileMasterMainAuth}
\title{GITHUB_PAT-aware and \code{main}-\code{master}-aware download from GitHub}
\usage{
.downloadFileMasterMainAuth(
  url,
  destfile,
  need = "HEAD",
  verbose = getOption("Require.verbose"),
  verboseLevel = 2
)
}
\value{
This is called for its side effect, namely, the same as \code{utils::download.file}, but
using a \code{GITHUB_PAT}, it if is in the environment, and trying both \code{master} and
\code{main} if the actual \code{url} specifies either \code{master} or \code{main} and it does not exist.
}
\description{
Equivalent to \code{utils::download.file}, but taking the \code{GITHUB_PAT} environment
variable and using it to access the Github url.
}
