% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDep.R
\name{.installed.pkgs}
\alias{.installed.pkgs}
\title{Partial alternative (faster) to \code{installed.packages}}
\usage{
.installed.pkgs(
  lib.loc = .libPaths(),
  which = c("Depends", "Imports", "LinkingTo"),
  other = NULL,
  purge = getOption("Require.purge", FALSE)
)
}
\description{
This reads the DESCRIPTION files only, so can only access fields that are
available in the DESCRIPTION file. This is different than \code{installed.packages}
which has many other fields, like "Built", "NeedsCompilation" etc. If those
fields are needed, then this function will return an empty colum in the returned
character matrix.
}
