% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envs.R
\name{envPkgCreate}
\alias{envPkgCreate}
\title{1st level --> create the .pkgEnv object in Require}
\usage{
envPkgCreate(parentEnv = asNamespace("Require"))
}
\description{
1st level --> create the .pkgEnv object in Require
}
