% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-helpers.R
\name{install_githubV}
\alias{install_githubV}
\title{Vectorized install_github}
\usage{
install_githubV(gitPkgNames, install_githubArgs, ...)
}
\arguments{
\item{gitPkgNames}{Character vector of package to install from github}

\item{install_githubArgs}{Any arguments passed to \code{install_github}}

\item{...}{Another way to pass arguments to \code{install_github}}
}
\description{
This will attempt to identify all dependencies of all packages first, then
load the packages in the correct order so that each of their dependencies are
met before each is installed.
}
\examples{
\dontrun{
  install_githubV(c("PredictiveEcology/Require", "PredictiveEcology/quickPlot"))
}

}
