% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDep3.R
\name{joinToAvailablePackages}
\alias{joinToAvailablePackages}
\title{Join a data.table with a \code{Package} column to \code{available.packages}}
\usage{
joinToAvailablePackages(pkgDT, repos, type, which, verbose)
}
\value{
The returned \code{data.table} will have most of the columns from
\code{available.packages} appended to the \code{pkgDT}, including \code{Depends}, \code{Imports},
\code{Suggests}. It will change the column name that is normally
returned from \code{available.packages} as \code{Version} to \code{VersionOnRepos}.
}
\description{
Will join \code{available.packages()} with \code{pkgDT}, if \code{pkgDT} doesn't already have
a column named \code{Depends}, which would be an indicator that this had already
happened.
}
