% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{linkOrCopy}
\alias{linkOrCopy}
\title{Create link to file, falling back to making a copy if linking fails.}
\usage{
linkOrCopy(to, from)
}
\arguments{
\item{from, to}{character vectors, containing file names or paths.
    For \code{file.copy} and \code{file.symlink}
#ifdef windows not active    \code{to} can alternatively be the path to a single existing directory.}
}
\description{
First try to create a hardlink to the file.
If that fails, try a symbolic link (symlink) before falling back to copying the file.
"File" here can mean a file or a directory.
}
