% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDep.R
\name{masterMainToHead}
\alias{masterMainToHead}
\title{This converts master or main to HEAD for a git repo}
\usage{
masterMainToHead(gitRepo)
}
\arguments{
\item{gitRepo}{A git repository of the form account/repo with optional @branch
or @sha or @tag}
}
\value{
The git repository with @HEAD if it had @master, @main or no @.
}
\description{
This will also convert a git repo with nothing after the @ to @HEAD
}
