% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require2.R
\name{mergeRemotes}
\alias{mergeRemotes}
\title{If a list of dependencies includes Remotes from DESCRIPTION file}
\usage{
mergeRemotes(depsList)
}
\description{
If a list of dependencies includes Remotes from DESCRIPTION file
}
