% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDep.R
\name{purgeCache}
\alias{purgeCache}
\title{Purge everything in the Require cache}
\usage{
purgeCache(repos = getOption("repos"))
}
\arguments{
\item{repos}{The remote repository (e.g., a CRAN mirror), passed to either
\code{install.packages}, \code{install_github} or \code{installVersions}.}
}
\value{
Run for its side effect, namely, all cached objects are removed.
}
\description{
Require uses caches for local Package saving, local caches of \code{available.packages},
local caches of GitHub (e.g., \code{"DESCRIPTION"}) files, and some function calls
that are cached. This function clears all of them.
}
