% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{R_TESTSomit}
\alias{R_TESTSomit}
\title{This environment variable "R_TESTS" is set during testing, and it points to a file
called \code{Startup.Rs} that is placed in the \code{.libPaths()}.
If the \code{.libPaths()} is changed during the testing,
then that file will not be found, and install.packages will
fail to install a package with an error of source file not found.
See: \url{https://github.com/HenrikBengtsson/startup/issues/19}.}
\usage{
R_TESTSomit()
}
\description{
The environment variable is set here:
\verb{https://github.com/wch/r-source/blob/8b6429feb661b02e2b2b6df1757b31cf1250a33e/src/library/tools/R/testing.R#L472-Lundefined}
}
\keyword{internal}
