% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-helpers.R
\name{rCurrentVersion}
\alias{rCurrentVersion}
\title{Get or compare current R version to a known version}
\usage{
rCurrentVersion(testVers)
}
\arguments{
\item{testVers}{A character string using format "== 4.1"
or ">= 4.1"}
}
\value{
If no \code{testVers} is supplied, then it will just return the current R version.
If \code{testVers} is supplied, then it will return a \code{TRUE} or \code{FALSE}.
}
\description{
Compares R version to a known version
}
\examples{
rCurrentVersion(">= 4.1")
}
