% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-package.R
\docType{package}
\name{Require-package}
\alias{Require}
\alias{Require-package}
\title{Require package}
\description{
A single key function, \code{Require} that wraps \code{install.packages},
\code{remotes::install_github}, \code{versions::install.versions}, and \code{base::require}
that allows for reproducible workflows. As with other functions in a reproducible workflow,
this package emphasizes functions that return the  same result whether it is the first or
subsequent times running the function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://Require.predictiveecology.org}
  \item \url{https://github.com/PredictiveEcology/Require}
  \item Report bugs at \url{https://github.com/PredictiveEcology/Require/issues}
}

}
\author{
\strong{Maintainer}: Eliot J B McIntire \email{eliot.mcintire@canada.ca} (\href{https://orcid.org/0000-0002-6914-8316}{ORCID})

Other contributors:
\itemize{
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
