% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{RequireCacheDir}
\alias{RequireCacheDir}
\alias{RequirePkgCacheDir}
\title{Path to (package) cache directory}
\usage{
RequireCacheDir()

RequirePkgCacheDir()
}
\description{
Path to (package) cache directory
}
