% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDep3.R
\name{rmBase}
\alias{rmBase}
\title{Recursive function to remove \code{.basePkgs}}
\usage{
rmBase(includeBase = formals(pkgDep)[["includeBase"]], deps)
}
\arguments{
\item{includeBase}{Logical. If \code{FALSE}, the default, then base packages will
be removed.}

\item{deps}{Either a list of dependencies, a data.table of dependencies with
a column \code{Package} or a vector of dependencies.}
}
\description{
Recursive function to remove \code{.basePkgs}
}
