% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{rpackageFolder}
\alias{rpackageFolder}
\title{Append the R version to a directory path}
\usage{
rpackageFolder(path = cacheGetOptionCachePkgDir(), exact = FALSE)
}
\arguments{
\item{path}{character specifying a directory path}

\item{exact}{logical indicating whether to use \code{path} as-is (i.e., without appending)}
}
\description{
Append the R version to a directory path
}
\keyword{internal}
