% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{rPkgDir}
\alias{rPkgDir}
\title{Append the R version to a directory path}
\usage{
rPkgDir(path = cacheGetOptionCachePkgDir(), exact = FALSE)
}
\arguments{
\item{path}{character specifying a directory path}

\item{exact}{logical indicating whether to use \code{path} as-is (i.e., without appending)}
}
\description{
Used primarily with \code{cacheGetOptionCachePkgDir}.
}
\keyword{internal}
