% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDep3.R
\name{saveNamesForCache3}
\alias{saveNamesForCache3}
\title{The \code{packages} argument may have up to 4 pieces of information for GitHub
packages: name, repository, branch, version. For CRAN-alikes, it will only
be 2 pieces: name, version. There can also be an inequality or equality, if
there is a version.}
\usage{
saveNamesForCache3(packages, which, recursive, ap, type = type, repos, verbose)
}
\arguments{
\item{packages}{A vector of GitHub or CRAN-alike packages. These can include
package name (required) and optionally repository, branch/sha and/or version.}

\item{ap}{is used for CRAN-alikes to get version number, if not part of \code{packages}}

\item{repos}{is used for \code{ap}.}
}
\value{
A (named) vector of SaveNames, which is a concatenation of the 2 or 4 elements
above, plus the \code{which} and the \code{recursive}.
}
\description{
The \code{packages} argument may have up to 4 pieces of information for GitHub
packages: name, repository, branch, version. For CRAN-alikes, it will only
be 2 pieces: name, version. There can also be an inequality or equality, if
there is a version.
}
\details{
If version is not supplied, it will take the local, installed version, if it
exists. Otherwise, it is assumed that the HEAD is desired.
The function will find it in the \code{ap} or on \code{github.com}. For github packages,
this is obviously a slow step, which can be accelerated if user supplies a sha
or a version e.g., saveNamesForCache("PredictiveEcology/LandR@development (==1.0.2)")
}
