% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setLinuxBinaryRepo}
\alias{setLinuxBinaryRepo}
\title{Setup for binary Linux repositories}
\usage{
setLinuxBinaryRepo(
  binaryLinux = urlForArchivedPkgs,
  backupCRAN = srcPackageURLOnCRAN
)
}
\arguments{
\item{binaryLinux}{A CRAN repository serving binary Linux packages.}

\item{backupCRAN}{If there is no CRAN repository set}
}
\value{
Called for its side effect on \code{options("repos")}; returns that \code{options()} result invisibly, or \code{NULL} when nothing needed changing.
}
\description{
Enable use of binary package builds for Linux from the RStudio Package
Manager repo. This will set the \code{repos} option, affecting the current R
session. It will put this \code{binaryLinux} in the first position. If the
\code{getOption("repos")} is \code{NULL}, it will put \code{backupCRAN} in second position.
}
