% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-helpers.R
\name{sourcePkgs}
\alias{sourcePkgs}
\title{A list of R packages that should likely be installed from Source, not Binary}
\usage{
sourcePkgs(additional = NULL, spatialPkgs = NULL, otherPkgs = NULL)
}
\arguments{
\item{additional}{Any other packages to be added to the other 2 argument vectors}

\item{spatialPkgs}{A character vector of package names that focus on spatial analyses.}

\item{otherPkgs}{A character vector of package names that often
require system specific compilation.}
}
\value{
A sorted concatenation of the 3 input parameters.
}
\description{
This list can be updated by the user by setting the value of \code{Require.spatialPkgs} or
\code{Require.otherPkgs} in the \code{options} or another environment. See \code{RequireOptions()}. If this function does
not find either of those two objects, then it will default to the (non exported)
ones in \code{Require}
}
