% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require2.R
\name{availablePackagesOverride}
\alias{availablePackagesOverride}
\title{Create a custom "available.packages" object}
\usage{
availablePackagesOverride(toInstall, repos, purge)
}
\description{
This is the mechanism by which \code{install.packages}
}
