% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{cacheGetOptionCachePkgDir}
\alias{cacheGetOptionCachePkgDir}
\title{Get the option for \code{Require.cachePkgDir}}
\usage{
cacheGetOptionCachePkgDir()
}
\description{
First checks if an environment variable \code{Require.cachePkgDir}
is set and defines a path.
If not set, checks whether the \code{options("Require.cachePkgDir")} is set.
If a character string, then it returns that.
If \code{TRUE}, then use \code{cachePkgDir()}. If \code{FALSE}
then returns \code{NULL}.
}
