% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDep.R
\name{cacheClearPackages}
\alias{cacheClearPackages}
\alias{clearRequirePackageCache}
\title{Clear Require Cache elements}
\usage{
cacheClearPackages(
  packages,
  ask = interactive(),
  Rversion = versionMajorMinor(),
  clearCranCache = FALSE,
  verbose = getOption("Require.verbose")
)

clearRequirePackageCache(
  packages,
  ask = interactive(),
  Rversion = versionMajorMinor(),
  clearCranCache = FALSE,
  verbose = getOption("Require.verbose")
)
}
\arguments{
\item{packages}{Either missing or a character vector of package names
(currently cannot specify version number) to remove from the local Require
Cache.}

\item{ask}{Logical. If \code{TRUE}, then it will ask user to confirm}

\item{Rversion}{An R version (major dot minor, e.g., "4.2"). Defaults to
current R version.}

\item{clearCranCache}{Logical. If \code{TRUE}, then this will also clear the
local \code{crancache} cache, which is only relevant if
\code{options(Require.useCranCache = TRUE)}, i.e., if \code{Require} is using the
\code{crancache} cache also}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, also returns details as if
\code{returnDetails = TRUE} (for backwards compatibility).}
}
\description{
Clear Require Cache elements
}
