% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require2.R
\name{compareVersion2}
\alias{compareVersion2}
\title{Compare package versions}
\usage{
compareVersion2(version, versionSpec, inequality)
}
\arguments{
\item{version}{One or more package versions. Can be \code{character} or \code{numeric_version}.}

\item{versionSpec}{One or more versions to compare to.
Can be \code{character} or \code{numeric_version}.}

\item{inequality}{The inequality to use, i.e., \code{>=}.}
}
\value{
a logical vector of the length of the longest of the 3 arguments.
}
\description{
Alternative to \code{utils::compareVersion} that is vectorized on \code{version},
\code{versionSpec} and/or \code{inequality}. This will also return an NA element
in the returned vector if one of the arguments has NA for that element.
}
