% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-helpers.R
\name{DESCRIPTIONFileVersionV}
\alias{DESCRIPTIONFileVersionV}
\alias{DESCRIPTIONFileOtherV}
\alias{getGitHubDESCRIPTION}
\title{GitHub package tools}
\usage{
DESCRIPTIONFileVersionV(file)

DESCRIPTIONFileOtherV(file, other = "RemoteSha")

getGitHubDESCRIPTION(pkg)
}
\arguments{
\item{file}{A file path to a DESCRIPTION file}

\item{other}{Any other keyword in a DESCRIPTION file that preceeds a ":". The rest of the line will be
retrieved.}

\item{pkg}{A character string with a GitHub package specification (c.f. remotes)}
}
\description{
A series of helpers to access and deal with GitHub packages
}
\details{
\code{getGitHubDESCRIPTION} retrieves the DESCRIPTION file from GitHub.com
}
