% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-helpers.R
\name{DESCRIPTIONFileVersion}
\alias{DESCRIPTIONFileVersion}
\alias{getGitHubDESCRIPTION}
\title{GitHub package tools}
\usage{
DESCRIPTIONFileVersion(file)

getGitHubDESCRIPTION(pkg)
}
\arguments{
\item{file}{A file path to a DESCRIPTION file}

\item{pkg}{A character string with a GitHub package specification (c.f. remotes)}
}
\description{
A series of helpers to access and deal with GitHub packages
}
\details{
\code{getGitHubDESCRIPTION} retrieves the DESCRIPTION file from GitHub.com
}
