% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractPkgName}
\alias{extractPkgName}
\alias{extractVersionNumber}
\alias{extractInequality}
\alias{extractPkgGitHub}
\title{Extract info from package character strings}
\usage{
extractPkgName(pkgs, filenames)

extractVersionNumber(pkgs, filenames)

extractInequality(pkgs)

extractPkgGitHub(pkgs)
}
\arguments{
\item{pkgs}{A character string vector of packages with or without GitHub path or versions}

\item{filenames}{Can be supplied instead of \code{pkgs} if it is a filename e.g., a
.tar.gz or .zip that was downloaded from CRAN.}
}
\value{
Just the package names without extraneous info.
}
\description{
Cleans a character vector of non-package name related information (e.g., version)
}
\examples{
extractPkgName("Require (>=0.0.1)")
extractVersionNumber(c(
  "Require (<=0.0.1)",
  "PredictiveEcology/Require@development (<=0.0.4)"
))
extractInequality("Require (<=0.0.1)")
extractPkgGitHub("PredictiveEcology/Require")
}
\seealso{
\code{\link[=trimVersionNumber]{trimVersionNumber()}}

Other version specifications:
\code{\link[=compareVersion2]{compareVersion2()}},
\code{\link[=trimRedundancies]{trimRedundancies()}},
\code{\link[=trimVersionNumber]{trimVersionNumber()}}
}
\concept{version specifications}
