% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-helpers.R
\name{extractInequality}
\alias{extractInequality}
\alias{extractPkgGitHub}
\alias{extractPkgName}
\title{Extract info from package character strings}
\usage{
extractInequality(pkgs)

extractPkgGitHub(pkgs)

extractPkgName(pkgs)
}
\arguments{
\item{pkgs}{A character string vector of packages with or without GitHub path or versions}
}
\value{
Just the package names without extraneous info.
}
\description{
Cleans a character vector of non-package name related information (e.g., version)
}
\examples{
extractPkgName("Require (>=0.0.1)")
}
