% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRAN.R
\name{getCRANrepos}
\alias{getCRANrepos}
\title{A helper function to get or set CRAN repos}
\usage{
getCRANrepos(repos = NULL, ind)
}
\arguments{
\item{repos}{A CRAN-like repository}

\item{ind}{an integer of which mirror to use in \code{chooseCRANmirror()}}
}
\description{
This will get the current option in \code{getOption('repos')}, and if that is not
set to a url, then it will prompt the user to select a mirror, unless
\code{ind} is set, in which case, it will use that mirror (in
\code{chooseCRANmirror()})
}
\keyword{internal}
