% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-helpers.R
\name{getOptionRPackageCache}
\alias{getOptionRPackageCache}
\title{Get the option for Require.RPackageCache}
\usage{
getOptionRPackageCache()
}
\description{
Takes the EnvVar first, then option as path, then option as logical; if logical then use RequirePkgCacheDir()
}
