% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{getOptionRPackageCache}
\alias{getOptionRPackageCache}
\title{Get the option for \code{Require.RPackageCache}}
\usage{
getOptionRPackageCache()
}
\description{
First checks if an environment variable \code{Require.RPackageCache}
is set and defines a path.
If not set, checks whether the \code{options("Require.RPackageCache")} is set.
If a character string, then it returns that.
If \code{TRUE}, then use \code{RequirePkgCacheDir()}. If \code{FALSE}
then returns \code{NULL}.
}
