% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require.R
\name{parseGitHub}
\alias{parseGitHub}
\alias{getGitHubDESCRIPTION}
\title{GitHub package tools}
\usage{
parseGitHub(pkgDT)

getGitHubDESCRIPTION(pkg)
}
\arguments{
\item{pkg}{A character string with a GitHub package specification (c.f. remotes)}
}
\value{
A data.table with added columns.
}
\description{
A series of helpers to access and deal with GitHub packages
}
\details{
\code{parseGitHub} turns the single character string representation into 3 or 4:
Account, Repo, Branch, SubFolder.

\code{getGitHubDESCRIPTION} retrieves the DESCRIPTION file from GitHub.com
}
