% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-helpers.R
\name{installGithubPackage}
\alias{installGithubPackage}
\alias{installGitHubPackage}
\title{Install R Package from GitHub source code}
\usage{
installGithubPackage(gitRepo, libPath = .libPaths()[1])

installGitHubPackage(gitRepo, libPath = .libPaths()[1])
}
\arguments{
\item{gitRepo}{A repository in the form: Account/Repository@Branch or Account/Repository@SHA}

\item{libPath}{The folder where you would like the package installed. Defaults
to \code{.libPaths()[1]}}
}
\description{
A lightweight alternative to \code{devtools::install_github}. All dependencies
must have been installed already for this to work.
}
