% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-helpers.R
\name{getPkgVersions}
\alias{getPkgVersions}
\title{Internals used by \code{Require}}
\usage{
getPkgVersions(pkgDT, install = TRUE, verbose = getOption("Require.verbose"))
}
\arguments{
\item{pkgDT}{A pkgDT data.table.}

\item{install}{Logical or "force". If \code{FALSE}, this will not try to install
anything. If \code{"force"}, then it will force installation of requested
packages, mimicking a call to e.g., \code{install.packages}. If \code{TRUE}, the
default, then this function will try to install any missing packages or
dependencies.}

\item{verbose}{Numeric or logical indicating how verbose should the function
be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE,
then minimal outputs; if \code{1} or TRUE, more outputs; \code{2} even more. NOTE: in
\code{Require} function, when \code{verbose >= 2}, the return object will have an
attribute: \code{attr(.., "Require")} which has lots of information about the
processes of the installs.}
}
\value{
In general, these functions return a \code{data.table} with various package
information, installation status, version, available version etc.
}
\description{
While these are not intended to be called manually by users, they may be
of some use for advanced users.
}
