% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{RequireCacheDir}
\alias{RequireCacheDir}
\alias{RequirePkgCacheDir}
\title{Path to (package) cache directory}
\usage{
RequireCacheDir(create = FALSE)

RequirePkgCacheDir(create = TRUE)
}
\arguments{
\item{create}{A logical indicating whether the path should
be created if it does not exist. Default is \code{FALSE}.}
}
\value{
If \code{create = TRUE}, the default behaviour, the cache directory will be created,
with a README placed in the folder. Otherwise, this function will just
return the path of what the cache directory would be.
}
\description{
Sets or gets the cache directory associated with the \code{Require} package.
}
