% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-helpers.R
\name{setLibPaths}
\alias{setLibPaths}
\title{Set .libPaths}
\usage{
setLibPaths(libPaths, standAlone = TRUE)
}
\arguments{
\item{libPaths}{A new path to append to, or replace all existing user
components of \code{.libPath()}}

\item{standAlone}{Logical. If \code{TRUE}, all packages will be installed to and loaded from
                  the \code{libPaths} only. If \code{FALSE}, then \code{libPath} will
                  be prepended to \code{.libPaths()} during the \code{Require} call,
                  resulting in shared packages, i.e.,
                  it will include the user's default package folder(s).
                  This can be create dramatically faster
installs if the user has a substantial number of the packages already in
their personal library. Default \code{FALSE} to minimize package installing.}
}
\description{
This will set the .libPaths() by either adding a new path to
it if \code{standAlone = FALSE}, or will concatenate
\code{c(libPath, tail(.libPaths(), 1))} if \code{standAlone = TRUE}.
}
\details{
This code was taken from \url{https://milesmcbain.xyz/hacking-r-library-paths/}
}
\examples{
\dontrun{
setLibPaths("~/newProjectLib") # will only have 2 paths
setLibPaths("~/newProjectLib", standAlone = FALSE) # will have 2 or more paths

}
}
