% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setLinuxBinaryRepo}
\alias{setLinuxBinaryRepo}
\title{Setup for binary Linux repositories}
\usage{
setLinuxBinaryRepo(binaryLinux = "https://packagemanager.rstudio.com/")
}
\arguments{
\item{binaryLinux}{A CRAN repository serving binary Linux packages.}
}
\description{
Enable use of binary package builds for Linux from the Rstudio Package Manager repo.
This will set the \code{repos} option, affecting the current R session.
}
