% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require-helpers.R
\name{sourcePkgs}
\alias{sourcePkgs}
\title{A list of R packages that should likely be installed from Source, not Binary}
\usage{
sourcePkgs(additional = NULL, spatialPkgs = NULL, otherPkgs = NULL)
}
\arguments{
\item{additional}{Any other packages to be added to the other 2 argument vectors}

\item{spatialPkgs}{A character vector of package names that focus on spatial analyses.}

\item{otherPkgs}{A character vector of package names that often
require system specific compilation.}
}
\value{
A sorted concatenation of the 3 input parameters.
}
\description{
The list of R packages that \code{Require} installs from source on Linux, even if
the \code{getOptions("repos")} is a binary repository. This list can be updated by
the user by modifying the options \code{Require.spatialPkgs} or
\code{Require.otherPkgs}. Default "force source only packages" are visible with
\code{RequireOptions()}.
}
