% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tempdir2}
\alias{tempdir2}
\title{Make a temporary (sub-)directory}
\usage{
tempdir2(
  sub = "",
  tempdir = getOption("Require.tempPath", .RequireTempPath()),
  create = TRUE
)
}
\arguments{
\item{sub}{Character string, length 1. Can be a result of
\code{file.path("smth", "smth2")} for nested temporary sub
directories.}

\item{tempdir}{Optional character string where the
temporary dir should be placed. Defaults to \code{.RequireTempPath()}}

\item{create}{Logical. Should the directory be created. Default \code{TRUE}}
}
\description{
Create a temporary subdirectory in \code{.RequireTempPath()}, or a
temporary file in that temporary subdirectory.
}
\seealso{
\code{\link[=tempfile2]{tempfile2()}}
}
