% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{trimVersionNumber}
\alias{trimVersionNumber}
\title{Trim version number off a compound package name}
\usage{
trimVersionNumber(pkgs)
}
\arguments{
\item{pkgs}{A character string vector of packages with or without GitHub path or versions}
}
\description{
The resulting string(s) will have only name (including github.com repository if it exists).
}
\examples{
trimVersionNumber("PredictiveEcology/Require (<=0.0.1)")
}
\seealso{
\code{\link[=extractPkgName]{extractPkgName()}}
}
